Add new explore and explore_api blueprints
